Rename ValidateJsonMiddleware.ts file.
